pyproject.toml
[project]
name = "a2sv-job-portal"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikiyas",email = "mikiyas.m.degefu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==4.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
